+++ matchup-tool/src/test/java/com/bc/fiduceo/matchup/AbstractUsecaseIntegrationTest.java
@@ -94,6 +94,8 @@ abstract class AbstractUsecaseIntegrationTest {
             satelliteObservation.setNodeType(acquisitionInfo.getNodeType());
             satelliteObservation.setVersion(version);
 
+            ReaderFactory.close();
+
             return satelliteObservation;
         }
     }